requirements.txt
playwright==1.42.0
pyotp==2.9.0
python-dotenv==1.0.1
supabase==2.3.6
psycopg2-binary==2.9.9
fastapi==0.110.0
uvicorn==0.28.0
